refactor: rename tombi-date-time to date-time and update paths
This commit renames the 'tombi-date-time' crate to 'date-time' and updates the corresponding paths in the Cargo files. It also ensures that dependencies and features are correctly defined in the new crate structure.
