+++ Cargo.toml
@@ -20,6 +20,7 @@ chrono = { version = "0.4.38", features = ["serde"] }
 clap = { version = "4.5.20", features = ["derive", "string"] }
 clap-verbosity-flag = "3.0.1"
 config = { path = "crates/config" }
+date-time = { path = "crates/date-time" }
 diagnostic = { path = "crates/diagnostic" }
 document = { path = "crates/document" }
 document-tree = { path = "crates/document-tree" }
