+++ crates/date-time/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "tombi-date-time"
+name = "date-time"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
@@ -10,12 +10,12 @@ license.workspace = true
 chrono = { workspace = true, optional = true }
 serde = { workspace = true, optional = true }
 
-[features]
-chrono = ["dep:chrono", "serde"]
-default = ["serde"]
-serde = ["dep:serde"]
-
 [dev-dependencies]
 pretty_assertions.workspace = true
 rstest.workspace = true
 serde_json.workspace = true
+
+[features]
+chrono = ["dep:chrono", "serde"]
+default = ["serde"]
+serde = ["dep:serde"]
