+++ crates/date-time/tests/serialize.rs
@@ -1,6 +1,6 @@
+use date_time::{LocalDate, LocalDateTime, LocalTime, Offset, OffsetDateTime};
 use rstest::rstest;
 use serde_json::json;
-use tombi_date_time::{LocalDate, LocalDateTime, LocalTime, Offset, OffsetDateTime};
 
 #[test]
 fn test_local_date_serialization() {
